:bulb: Move error to util/ and separate hex formatting
